Add time-of-day mode
Add support for a mode (requested via the #mode=tod option) in which the
only display is a time-of-day clock with the date beneath.  This can be
used to create a plain clock display which happens to be driven by the
official timing clock.

Future project: remember the offset between local time and the rmonitor
feed time-of-day, and use this to display a free-running-but-correct TOD
clock when there is no server connection.