client: improve connection-lost message display
On the race info screen, move the leader/TOD display down slightly to
make room for the "Connection Lost" message to appear on a separate
line.  Instead of displaying the message by changing its content
programmatically, leave the content fixed and toggle the element's
visibility.  This prevents the TOD display from moving when the message
is displayed or removed, and leaves a small gap above the leader/TOD
line, resulting in a more pleasing display in normal operation.