client: normalize mode selection
Instead of the options-processing code having to know about every
possible display mode, define an array of modes and make sure they are
all structurally similar.  Particularly, every mode is shown in a single
element (a div) whose ID is the same as the mode name and which expects
to be displayed in block mode when visible.

Note this changes the name of the ToD mode from "tod" to "bigtod", and
using a mode other than "bigtod" or "raceinfo" will result in a blank
screen.  The default mode remains "raceinfo".