Handle style choices for different monitors
This allows selection from a number of different configurations for
different displays, by means of a new "display" option.  The value of
this option, if given, is added to the classes of a new top-level div
element.  This effectively allows different styles to be selected for
different monitor configurations.

Two display configurations are added for use on Chromecast devices
displaying to 16:9 displays (chromeHD) and 4:3 displays (chrome43).
These options configure the height of the top-level div, add a 50-pixel
top margin, and limit the width based on the available display area
(particularly on 4:3 devices, where the full width cannot be used
because the browser does not know that the sides are cut off).

The default configuration uses the full available display area, and is
suitable for use in a normal browser.