server: WebSocketServer: client list thread safety
Improve thread safety for things that enumerate the client list, by
iterating over the actual set of clients instead of iterating over the
list of keys and assuming each key is actually still in use by the time
we get to it.

In the case of the poller thread, this also improves thread safety by
getting us our own reference to a client and using it consistently,
instead of repeatedly looking it up in the client list.  Note, however,
that readData is still not thread-safe, and the current poller design
will not work correctly unless clients only ever send pong messages
_and_ the poller is the only thing reading them.

An artifact of this change is that clients now have an 'id' field which
contains the client ID (connection hash), which is also the key used for
that client in the client list.