server: WebSocketServer: avoid lock on remove
Avoid unneeded locking in _remove() and close() when removing a client.
We don't need a lock just to remove a client from the client list, and
there's no point in synchronizing that removal with closing the client
socket.  Of course, things would be different if there were threads that
counted on the client list not changing over a period of time, but in
fact no such threads exist.  The closest is the interaction between the
listener and poller threads, which depend on synchronization only with
each other, to avoid getting into a state with a client connected but no
poller running.