server: relay: handle StopIteration
When the Orbits server shuts down normally, it closes all rmonitor
connections.  This looks like an EOF on the socket, so reading the next
line results in a StopIteration exception instead of an error.  Handle
that case the same as if we had lost the connection due to an error.