fix timezone issues on some clients (chromecast)
Add new mechanisms which allow the client's timezone to be set via a
fragment option ("tz=TIMEZONE") or a server message (:TZ), with the
former overriding the latter.  The server sends this message when a new
client connects, during a refresh, and any time it is changed.  For now,
the server's timezone defaults to America/New_York and can only be
changed via a console command.