Make data port configurable, with default 9999
Allow the data port to be set via a server configuration option.
The port number is passed to the javascript client code by means of a
cookie returned by our HTTP server.

Additionally, the default port is now 9999 rather than 9876, to avoid a
conflict with SSH tunnel configuration on some of my machines.