server: clean up console output
Introduce a new local reply() function in ws.consoleCommand(), which
simplifies sending replies to the client that submitted a command.
In the process, we also ensure that there is a blank line after the
output of each command, which improves usability of the console in the
absence of a command prompt.

Finally, fix a bug in which the unrecognized command error was sent to
all clients instead of only the console that submitted the bad command.