server: management gui frame
Add the beginnings of a management GUI.  This makes the changes needed
to support running a Tk-based GUI in the main thread and still have
everything else work.  A new command-line option is added to request GUI
mode.

The GUI itself is dirt-simple -- it consists only of a "Quit" button.