Ignore cars not started
When computing overall and class leaders, ignore cars with zero total
time, indicating they have not been seen since the green flag. Such a
car cannot be a leader; they probably aren't even on the track!

The exception is the car in overall first place, which will have zero
time during the first lap of a race started on the first passing.