rmonitor.RMonitorCollector: connect -> pull_once
Rename RMonitorCollector.connect() to pull_once(), to make it more clear
that it is a one-shot version of the loop that start_pull() kicks off
(in fact, that loop is basically just repeated calls to pull_once()).
Also, adopt the notion that the host and port arguments to both of these
can be given positionally, and start doing so.