server: avoid stderr output in Windows GUI mode
When invoked on Windows via explorer, or any other way that doesn't
connect us to a terminal, py2exe tries to redirect stderr to a log,
which results in permissions problems. Instead, redirect to the same
black hole as stdout. Since it's not easy to test for "connected to a
terminal", we instead check whether we are running in GUI mode.

This is a partial workaround. A more complete solution would arrange,
in non-GUI mode, for stderr to be sent to the terminal instead of to a
log file.