server: fix paned-window child lookup
The panes() method of a PanedWindow can apparently, depending on the
platform, return either a list of the child widgets or a list of their
names. Instead of looking for a "string" attribute, which doesn't exist
on the latter, call the str() method, which is the documented way to get
the path name of a widget and, of course, also works on things that are
already strings.